fix # 3
