+++ api/src/main/proto/input/reqreply/chat/chatservices.proto
@@ -5,12 +5,10 @@ import "input/reqreply/chat/findOne.proto";
 
 package rys.ajaxpetproject.service.chat;
 
-service ChatFindAllService {
+service ChatService {
   rpc ChatFindAll(request.findOne.create.proto.ChatFindOneRequest)
       returns (stream request.findOne.create.proto.ChatFindOneResponse);
-}
 
-service ChatCreateService {
   rpc createChat(request.chat.create.proto.ChatCreateRequest)
       returns (request.chat.create.proto.ChatCreateResponse);
 }
